# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddProductRequest:
    properties:
      cart_id:
        type: string
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - cart_id
    - product_id
    - quantity
    type: object
  dto.Cart:
    properties:
      id:
        type: string
      lines:
        items:
          $ref: '#/definitions/dto.CartLineRequest'
        type: array
      user:
        $ref: '#/definitions/internals_cart_controller_dto.User'
    type: object
  dto.CartLineRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.ListOrdersResponse:
    properties:
      orders:
        items:
          $ref: '#/definitions/dto.Order'
        type: array
      pagination:
        $ref: '#/definitions/paging.Pagination'
    type: object
  dto.Order:
    properties:
      code:
        type: string
      id:
        type: string
      lines:
        items:
          $ref: '#/definitions/dto.OrderLine'
        type: array
      status:
        type: string
      total_price:
        type: number
    type: object
  dto.OrderLine:
    properties:
      price:
        type: number
      product:
        $ref: '#/definitions/internals_order_controller_dto.Product'
      quantity:
        type: integer
    type: object
  dto.PlaceOrderLineRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
  dto.PlaceOrderRequest:
    properties:
      lines:
        items:
          $ref: '#/definitions/dto.PlaceOrderLineRequest'
        maxItems: 5
        type: array
      user_id:
        type: string
    required:
    - lines
    - user_id
    type: object
  dto.RemoveProductRequest:
    properties:
      cart_id:
        type: string
      product_id:
        type: string
    required:
    - cart_id
    - product_id
    type: object
...
  internals_cart_controller_dto.User:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
  internals_order_controller_dto.Product:
    properties:
      code:
        type: string
      id:
        type: string
      name:
        type: string
      price:
        type: number
    type: object
...
  paging.Pagination:
    properties:
      has_next:
        type: boolean
      has_previous:
        type: boolean
      page:
        type: integer
      size:
        type: integer
      skip:
        type: integer
      take_all:
        type: boolean
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
    type: object
host: localhost:8080
info:
  contact:
    email: anhquoc18092003@gmail.com
    name: Tran Phuoc Anh Quoc
  description: Swagger API for Go Clean Architecture.
  license:
    name: MIT
    url: https://github.com/MartinHeinz/go-project-blueprint/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Ecommerce Clean Architecture Swagger API
  version: "1.0"
paths:
...
  /carts/{userID}:
    get:
      description: Fetches the shopping cart details of the authenticated user based
        on the provided user ID.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved the user's cart
          schema:
            $ref: '#/definitions/dto.Cart'
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Cart not found for the given user ID
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Retrieve the cart of a user
      tags:
      - Carts
  /carts/{userID}/products:
    delete:
      consumes:
      - application/json
      description: Removes a specified product from the authenticated user's shopping
        cart.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product details to remove from cart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RemoveProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Remove product from cart successfully
          schema:
            type: string
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Remove a product from the user's cart
      tags:
      - Carts
    post:
      consumes:
      - application/json
      description: Adds a specified product to the authenticated user's shopping cart.
      parameters:
      - description: User ID
        in: path
        name: userID
        required: true
        type: string
      - description: Product details to add to cart
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Add product to cart successfully
          schema:
            type: string
        "400":
          description: Bad Request - Invalid request parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User ID mismatch or authentication failed
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add a product to the user's cart
      tags:
      - Carts
  /orders:
    get:
      description: Retrieve a list of orders for the authenticated user with optional
        filters.
      parameters:
      - description: Filter by order code
        in: query
        name: code
        type: string
      - description: Filter by order status
        in: query
        name: status
        type: string
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Number of records per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Field to order by (e.g., created_at)
        in: query
        name: order_by
        type: string
      - description: 'Sort order: true for descending, false for ascending'
        in: query
        name: order_desc
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Orders retrieved successfully
          schema:
            $ref: '#/definitions/dto.ListOrdersResponse'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get my orders
      tags:
      - Orders
    post:
      description: Creates a new order for the authenticated user.
      parameters:
      - description: Order details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.PlaceOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Order placed successfully
          schema:
            $ref: '#/definitions/dto.Order'
        "400":
          description: Bad Request - Invalid parameters
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden - User does not have the required permissions
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Place a new order
      tags:
      - Orders
  /orders/{id}:
    get:
      description: Retrieve details of a specific order by its ID.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order retrieved successfully
          schema:
            $ref: '#/definitions/dto.Order'
        "400":
          description: Bad Request - Missing or invalid Order ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Order does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Get order details
      tags:
      - Orders
  /orders/{id}/{status}:
    put:
      description: Update the status of an existing order.
      parameters:
      - description: Order ID
        in: path
        name: id
        required: true
        type: string
      - description: New order status
        in: path
        name: status
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order updated successfully
          schema:
            $ref: '#/definitions/dto.Order'
        "400":
          description: Bad Request - Missing or invalid Order ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized - User not authenticated
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found - Order does not exist
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error - An error occurred while processing
            the request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      - ApiKeyAuth: []
      summary: Update order status
      tags:
      - Orders
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"